requirements.txt



# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9
pydantic==2.9.2

# AI & Embeddings (OPTIMIZED for App Engine)
openai==1.55.3
tiktoken==0.7.0
sentence-transformers==3.3.1
torch==2.3.0
transformers==4.46.3

# Vector DB
qdrant-client==1.15.1

# Cloud
google-cloud-storage==2.18.2

# Web
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utils
python-dotenv==1.0.0
instructor==1.6.3

# Testing (optional - can remove to save space)
pytest==8.3.3
httpx==0.27.2

# Workflow & Graph
langgraph==0.2.28